requirements.txt
bottle==0.10.2
python-dateutil==1.5
Markdown==2.1.0
Jinja2==2.6
CherryPy==3.2.2
dictshield==0.4.3
python-memcached==1.48